requirements.txt
mysql-connector-python
pandas
numpy
scikit-learn
scikit-surprise
